refactor: extract donut helpers to alias-free module for testability
Move contextPercent, donutColor, donutStroke, formatTokenCount into
context-donut-utils.ts with a relative import to lib/types. Tests now
import from the utils module and actually run under bun test (previously
they crashed on @/ alias resolution).
